[prepifg] simpler python resample code and associated prepifg changes

+++ pyrate/tests/test_gdal_python.py
@@ -42,26 +42,44 @@ class TestResample(unittest.TestCase):
 
         for res in resolutions:
             res = [res, -res]
-            cmd = ['gdalwarp', '-overwrite', '-srcnodata', 'None', '-q', '-te']\
-                  + extents_str + ['-tr']
-            new_res_str = [str(r) for r in res]
-            cmd += new_res_str
-            for s in sydney_test_ifgs:
-                temp_tif = tempfile.mktemp(suffix='.tif')
-                t_cmd = cmd + [s.data_path,  temp_tif]
-                subprocess.check_call(t_cmd)
-                resampled_ds = gdal.Open(temp_tif)
-                resampled_ref = resampled_ds.ReadAsArray()
+            self.check_same_resampled_output(extents, extents_str, res,
+                                             sydney_test_ifgs)
 
-                rast = gdal.Open(s.data_path)
-                resampled_temp_tif = tempfile.mktemp(suffix='.tif',
-                                                    prefix='resampled_')
-                resampled = gdalwarp.resample(s.data_path, extents, res,
-                                                    resampled_temp_tif)
-                np.testing.assert_array_almost_equal(resampled_ref, resampled)
-                rast = None  # manual close
-                os.remove(temp_tif)
-                os.remove(resampled_temp_tif)  # this also proves file created
+    def check_same_resampled_output(self, extents, extents_str, res,
+                                    sydney_test_ifgs):
+        cmd = ['gdalwarp', '-overwrite', '-srcnodata', 'None', '-q', '-te'] \
+              + extents_str
+
+        if res[0]:
+            new_res_str = [str(r) for r in res]
+            cmd += ['-tr'] + new_res_str
+        for s in sydney_test_ifgs:
+            temp_tif = tempfile.mktemp(suffix='.tif')
+            t_cmd = cmd + [s.data_path, temp_tif]
+            subprocess.check_call(t_cmd)
+            resampled_ds = gdal.Open(temp_tif)
+            resampled_ref = resampled_ds.ReadAsArray()
+
+            rast = gdal.Open(s.data_path)
+            resampled_temp_tif = tempfile.mktemp(suffix='.tif',
+                                                 prefix='resampled_')
+            resampled = gdalwarp.resample(s.data_path, extents, res,
+                                          resampled_temp_tif)
+            np.testing.assert_array_almost_equal(resampled_ref, resampled)
+            rast = None  # manual close
+            os.remove(temp_tif)
+            if os.path.exists(resampled_temp_tif):
+                os.remove(resampled_temp_tif)
+
+    def test_none_resolution_output(self):
+        sydney_test_ifgs = common.sydney_data_setup()
+        # minX, minY, maxX, maxY = extents
+        extents = [150.91, -34.229999976, 150.949166651, -34.17]
+        extents_str = [str(e) for e in extents]
+
+        self.check_same_resampled_output(extents, extents_str, [None, None],
+                                         sydney_test_ifgs)
+
 
     def test_output_file_written(self):
         sydney_test_ifgs = common.sydney_data_setup()
